+++ NEWS.md
@@ -1,9 +1,10 @@
 ## 2024-11-15 3.10.0-SNAPSHOT
 * [MODSOURMAN-1249](https://folio-org.atlassian.net/browse/MODSOURMAN-1249) Added DI_MARC_FOR_UPDATE_RECEIVED log message handling
 * [MODSOURMAN-1266](https://folio-org.atlassian.net/browse/MODSOURMAN-1266) Event DI_JOB_COMPLETED is not being sent upon the completion of the data import process on snapshot, snapshot-2
-* [MODSOURMAN-1276](https://folio-org.atlassian.net/browse/MODSOURMAN-1276) Additional 147/447/547 to authority mappings
+* [MODSOURMAN-1276](https://folio-org.atlassian.net/browse/MODSOURMAN-1276) Add MARC fields 147/447/547 to authority mappings
 * [MODSOURMAN-1281](https://folio-org.atlassian.net/browse/MODSOURMAN-1281) 500 Error Returned when Posting Records Chunk with Existing ID Value
 * [MODSOURMAN-1282](https://folio-org.atlassian.net/browse/MODSOURMAN-1282) Marc Authority DI completed, but status changed fo failed
+* [MODSOURMAN-1288](https://folio-org.atlassian.net/browse/MODSOURMAN-1288) Add MARC fields 180/480/580 to authority mappings
 
 ## 2024-10-29 v3.9.0
 * [MODSOURMAN-1232](https://folio-org.atlassian.net/browse/MODSOURMAN-1232) Add the option to exclude job profile names to GET "/metadata-provider/jobExecutions" endpoint